Create Week 207
This was my first LeetCode contest, I only had time to do the easy problem but it was still fun! 
It's a mess because I was trying to finish as quickly as possible.
I submitted but then there were 2 test cases that didn't work so I had to come up with something to accommodate them.
Would definitely participate again!

The instructions:
You are given a string text of words that are placed among some number of spaces. Each word consists of one or more lowercase English letters and are separated by at least one space. It's guaranteed that text contains at least one word.

Rearrange the spaces so that there is an equal number of spaces between every pair of adjacent words and that number is maximized. If you cannot redistribute all the spaces equally, place the extra spaces at the end, meaning the returned string should be the same length as text.
Source: https://leetcode.com/contest/weekly-contest-207/problems/rearrange-spaces-between-words/