Create 9_15_20
I completed today's Leetcode challenge "Length of Last Word". The instructions: Given a string s consists of upper/lower-case alphabets and empty space characters ' ', return the length of last word (last word means the last appearing word if we loop from left to right) in the string.

If the last word does not exist, return 0.

Important concepts reviewed: Strings are immutable; Recursion

One of the test cases that I failed on was when the String was "a ". The method I made deleted characters prior to a space, so for that test case it deleted the whole String, leaving it empty. This meant I had to make a method that would eliminate a space at the end. I decided to use the trim() method. Here's what I thought would work:

    public static String removeEndSpace(String s) {
        return s.trim();
    } 

Which would be called by removeEndSpace(s); The problem is, that doesn't actually change the original String. You can't change a String, you'd have to make a new one. Therefore I was able to fix it by calling it as

	String oneWord = removeEndSpace(s);

After which I was able to pass oneWord into my removeWords method.

Recursion is something I still need help with, so I decided to try and make a method to remove words recursively. If the String has no spaces, meaning it was just a single word, base case reached! Otherwise, it would create a substring starting at the index after the space, and just keep chopping off words.

    public static String removeWords(String s) {
       	if(!s.contains(" ")) {
	    return s;
	} else {
	    String s1 = s.substring(s.indexOf(" ")+1);
	    return(removeWords(s1));
	}
    }

I know there has to be a faster method but I thought this would be fun.