Create Round_F_2021_Trash_Bins.java
Input: a String made of 0's and 1's, and an integer equal to the length of the string
Output: the sum of the amount of houses each household needs to walk past to reach a trash bin
Note: there will be at least 1 trash bin; no input would consist of only 0's or other characters

The premise of this problem is to imagine a street of houses, some with trash bins and some without. Houses with bins are marked 1 and those without are 0. If someone lives in a house with a bin, they have 0 houses they need to pass. But someone without would need to find their nearest bin. It's possible that they can only walk to the right or only walk to the left, or they might have to see if one direction has a bin closer than the other.

What I've learned from this problem:
-Look to see where your initialized variables are before submitting! It's so tempting to just run without checking. During the timed mode, I could have passed the first test set if I hadn't misplaced the answer variable (such a simple mistake, the answer was outside the loop). My initial solution passed the sample solution because it didn't make a difference for the tiny sample solution, so I didn't even think the location of the answer variable was wrong.

-Check the constraints. It's possible your ints will actually need to be longs. When I finally got a solution that was fast enough, I eagerly submitted without making sure that the oneTrashBin method should have been changed from int to long. I wasted attempts here! (Not that it matters per se, since this is just for practice)

-Come up with as many test cases as possible! In the competitive round, you aren't given the test cases, so you have to think of them on your own. Some scenarios and edge cases that you should have tested for in this problem are:
     -Only 1 bin (at the first house, somewhere in the middle, at the last house)
     -One bin at the first and last house
     -Streets that are REALLY long
     -Every house having a trash bin
     -A street with only 1 house

-Time Limit Exceeded: Before going to work coding, look at the pseudocode and think if there's a more time efficient solution. My initial algorithm was to go through every house, looking to the left and right, and just take the smaller of the two (if there were two) options. After getting a time limit exceeded with the larger test set, I realized there's a much better solution. For example, if the street started out looking like 0000100.... the first 4 residents are going to have to go to the 5th house for their nearest bin. There's no need to search for the nearest bin with every single resident! This way I devised a method where I could keep track of each bin to the left and to the right, and watching out for the edge cases when there was only one direction they could go.