Create Simple_Bank_System.java
What I learned: If it passes the sample but not the bigger test cases, its not necessarily a problem with your algorithm, or constructors.
Small typing errors can cause the code to compile, but not provide accurate results!

Tip for the next time: step away from the code and come back, reread everything like you would if you were proofreading a paper. In my case, what was tripping me up, was I had a =+ instead of += which was just a typo but made my code fail.