Create Round_B_2021_Increasing_Substring.java
For this problem, you were given a string and its length, and needed to print out, for each letter, the longest increasing substring ending on that letter. I realized every string is going to start with 1. There were no empty strings, and the strings were made up of capital letters. At first I thought I might have to be tracking the longest substring in the whole word, but nope, it was simpler than that. It was just a matter of making sure it printed out correctly (each test on its own line, a space before the number after printing the first "1")..