Create MergeSortedArray.java
I decided to revisit this problem after a friend came across it in an interview.
I had previously done this problem last August but did a really lazy solution by using Arrays.sort().  
This time I wanted to see if I could make it O(m+n) time.

For this problem, you are given two sorted arrays and you need to merge them into one sorted array. I wanted to accomplish this without making a new array (that would be too simple!). Because you can't change the length of an array, the first array is filled in with 0's where the extra numbers would go. You're also given the lengths of the two arrays (not counting any 0 placeholders).

After playing around with it on scratch paper, I decided to just write the values into the first array, starting with the largest, and working my way down to the beginning. I'd have to deal with the array going out of bounds when I decremented left_ptr and right_ptr so I made flags. This worked but I also had to consider the edge cases, what if an array came empty? So for that I made a switch statement since there were 3 cases. And this ended up being much faster than my solution from last year!