Create RomanToInteger.java
Initially I was going to look for the special cases and remove those letters from the string down and THEN add up the regular cases, but I saw that it would be a lot more straightforward to just subtract if there were any special cases.
This problem assumed that the given string was a valid roman numeral. If that wasn't a constraint, I would have initially gone through the string to make sure it only contained the 7 acceptable roman numeral letters. Also the range of answers was 1-3999 so I didn't have to worry about numbers being negative or too large.