Create intToRoman.java
Since I did the reverse problem previously I just had to do this one. It was a bit tricky getting the right values figured out, especially for the larger numbers but I ended up just trial and error until they all computed correctly. If I had to do a similar problem, a switch statement might be a bit more aesthetically pleasing instead of a bunch of if statements.