Create climbStairs.java
In honor of my university having a ton of stairs, I just had to do this problem.
After writing it out, I realized they were looking for a Fibonacci pattern.
The problem was, with a stair step of 45 steps, it took too long to calculate. My computer took literal seconds. A much quicker solution was to make an array and access the element at that step.