Create ValidParentheses.py
I used a stack to keep track of the left half of parentheses that have already been seen. When a right one appeared, I checked to see if the top of the stack was a mate, checking with the dictionary (which seemed cleaner than comparing with a bunch of ifs or switch statements). If you got through the whole string, for it to be valid, there had to be a final check to see if the stack was empty. Only then would this return true.