Create ReplaceElementsinArray.py
This problem was a challenge because my first solution passed the initial testcases but then I got time limit exceeded. Afterwards I realized that searching for the value in the array was too slow. A dictionary would be faster because accessing a value is constant time. Then I could just update it. So I'll keep that in mind, use a dictionary to search for values in an array (after you transfer the keys:values -> elements:indices).