Create SubstrWithoutRepeatingChar.cpp
This problem tested my ability to properly use queues and stacks. I was going for a sliding window technique with the queue, so I could add and remove letters within the window. But I also had to keep track of the letters being used with a stack so that it could remove not just 1 letter at a time, but all up to the duplicated letter.